# demo/message-center-demo/contracts/openapi/message-center-v1.yaml
openapi: 3.1.0
info:
  title: CRM for Logistics Message Center API
  version: 1.0.0
  description: Versioned HTTP contract for the message center frontend and backend.
  contact:
    name: CRM for Logistics Engineering
servers:
  - url: /
...
security:
  - sessionCookie: []
paths:
...
  /api/admin/whatsapp/cams/{scopeId}/callbacks:
    get:
      operationId: getAdminWhatsAppCallbacks
      description: Return the administrator-only desired CAMS callback configuration and latest write result. Provider state remains UNKNOWN because CAMS has no confirmed official readback API. Secrets, CustSpaceId, full phone numbers, and provider request identifiers are never returned.
      tags:
        - Admin
      x-required-role: ADMIN
      parameters:
        - $ref: '#/components/parameters/TraceRequestId'
        - $ref: '#/components/parameters/CamsScopeId'
      responses:
        '200':
          description: Redacted callback configuration projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWhatsAppCallbackConfig'
        '4XX':
          $ref: '#/components/responses/ApiErrorResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
  /api/admin/whatsapp/cams/{scopeId}/callbacks/phones/{channelAccountId}:
    put:
      operationId: updateAdminWhatsAppPhoneCallback
      description: Write upstream-message and delivery-status callback URLs for one phone through CAMS UpdatePhoneWebhook. The server resolves provider identity and credentials; the browser supplies only URLs, flags, and the optimistic version.
      tags:
        - Admin
      x-required-role: ADMIN
      parameters:
        - $ref: '#/components/parameters/TraceRequestId'
        - $ref: '#/components/parameters/CsrfTokenHeader'
        - $ref: '#/components/parameters/CamsScopeId'
        - $ref: '#/components/parameters/ChannelAccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminWhatsAppPhoneCallbackRequest'
      responses:
        '200':
          description: Redacted callback projection after a successful provider write
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWhatsAppCallbackConfig'
        '4XX':
          $ref: '#/components/responses/ApiErrorResponse'
        '409':
          $ref: '#/components/responses/ApiErrorResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
  /api/admin/whatsapp/cams/{scopeId}/callbacks/account:
    put:
      operationId: updateAdminWhatsAppAccountCallback
      description: Write the account-level delivery-status callback through CAMS UpdateAccountWebhook. Account-level requests cannot carry an upstream-message URL, provider identity, or credentials.
      tags:
        - Admin
      x-required-role: ADMIN
      parameters:
        - $ref: '#/components/parameters/TraceRequestId'
        - $ref: '#/components/parameters/CsrfTokenHeader'
        - $ref: '#/components/parameters/CamsScopeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminWhatsAppAccountCallbackRequest'
      responses:
        '200':
          description: Redacted callback projection after a successful provider write
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminWhatsAppCallbackConfig'
        '4XX':
          $ref: '#/components/responses/ApiErrorResponse'
        '409':
          $ref: '#/components/responses/ApiErrorResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: MESSAGE_CENTER_SESSION
      description: Opaque session token. The server stores only its SHA-256 digest.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: Header signature scheme. The adapter also requires X-Webhook-Timestamp and X-Webhook-Nonce for applicable channel types and verifies the exact raw body before parsing.
    weComSignature:
      type: apiKey
      in: query
      name: msg_signature
      description: WeCom query signature scheme. The adapter also requires timestamp and nonce query parameters and verifies the exact raw XML body before parsing.
  parameters:
    TraceRequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
    CsrfTokenHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 512
    CamsScopeId:
      name: scopeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ChannelAccountId:
      name: channelAccountId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ApiErrorResponse:
      description: Structured failure with no credential or signed URL material
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    AdminWhatsAppCallbackConfig:
      type: object
      additionalProperties: false
      required: [scopeId, phoneConfigs, accountConfig, ingressPath]
      description: Redacted desired configuration and latest write result. It never includes credentials, CustSpaceId, full phone numbers, raw provider responses, or provider request identifiers.
      properties:
        scopeId:
          type: string
          format: uuid
        phoneConfigs:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/AdminWhatsAppPhoneCallbackConfig'
        accountConfig:
          oneOf:
            - $ref: '#/components/schemas/AdminWhatsAppAccountCallbackConfig'
            - type: 'null'
        ingressPath:
          type: string
          const: /api/v1/webhooks/chatapp
    AdminWhatsAppPhoneCallbackConfig:
      type: object
      additionalProperties: false
      required: [channelAccountId, maskedPhone, desiredUpCallbackUrl, desiredStatusCallbackUrl,
        httpFlag, queueFlag, providerState, lastApplyStatus, lastErrorCode, lastAppliedAt, version]
      properties:
        channelAccountId:
          type: string
          format: uuid
        maskedPhone:
          type: string
          maxLength: 32
        desiredUpCallbackUrl:
          type: [string, 'null']
          maxLength: 2048
        desiredStatusCallbackUrl:
          type: [string, 'null']
          maxLength: 2048
        httpFlag:
          $ref: '#/components/schemas/WhatsAppCallbackFlag'
        queueFlag:
          $ref: '#/components/schemas/WhatsAppCallbackFlag'
        providerState:
          type: string
          enum: [UNKNOWN]
        lastApplyStatus:
          $ref: '#/components/schemas/WhatsAppCallbackApplyStatus'
        lastErrorCode:
          type: [string, 'null']
          maxLength: 100
        lastAppliedAt:
          type: [string, 'null']
          format: date-time
        version:
          type: integer
          format: int64
          minimum: 0
    AdminWhatsAppAccountCallbackConfig:
      type: object
      additionalProperties: false
      required: [desiredStatusCallbackUrl, httpFlag, queueFlag, providerState,
        lastApplyStatus, lastErrorCode, lastAppliedAt, version]
      properties:
        desiredStatusCallbackUrl:
          type: [string, 'null']
          maxLength: 2048
        httpFlag:
          $ref: '#/components/schemas/WhatsAppCallbackFlag'
        queueFlag:
          $ref: '#/components/schemas/WhatsAppCallbackFlag'
        providerState:
          type: string
          enum: [UNKNOWN]
        lastApplyStatus:
          $ref: '#/components/schemas/WhatsAppCallbackApplyStatus'
        lastErrorCode:
          type: [string, 'null']
          maxLength: 100
        lastAppliedAt:
          type: [string, 'null']
          format: date-time
        version:
          type: integer
          format: int64
          minimum: 0
    AdminWhatsAppPhoneCallbackRequest:
      type: object
      additionalProperties: false
      required: [upCallbackUrl, statusCallbackUrl, httpFlag, queueFlag, expectedVersion]
      properties:
        upCallbackUrl:
          type: string
          maxLength: 2048
          description: Empty clears the desired URL; non-empty values must use HTTPS.
        statusCallbackUrl:
          type: string
          maxLength: 2048
          description: Empty clears the desired URL; non-empty values must use HTTPS.
        httpFlag:
          $ref: '#/components/schemas/WhatsAppCallbackFlag'
        queueFlag:
          $ref: '#/components/schemas/WhatsAppCallbackFlag'
        expectedVersion:
          type: integer
          format: int64
          minimum: 0
    AdminWhatsAppAccountCallbackRequest:
      type: object
      additionalProperties: false
      required: [statusCallbackUrl, httpFlag, queueFlag, expectedVersion]
      properties:
        statusCallbackUrl:
          type: string
          maxLength: 2048
          description: Empty clears the desired URL; non-empty values must use HTTPS.
        httpFlag:
          $ref: '#/components/schemas/WhatsAppCallbackFlag'
        queueFlag:
          $ref: '#/components/schemas/WhatsAppCallbackFlag'
        expectedVersion:
          type: integer
          format: int64
          minimum: 0
    WhatsAppCallbackFlag:
      type: string
      enum: [Y, N]
    WhatsAppCallbackApplyStatus:
      type: string
      enum: [NEVER_APPLIED, SUCCEEDED, FAILED]
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, traceId, fieldErrors]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 1000
        traceId:
          type: string
          minLength: 1
          maxLength: 128
        fieldErrors:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 200
        code:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 500
...